Split name mempool code into separate file.
This is a backport of 1d9535cff103506c99e8147e6673cbf1819d6556, which
itself is basically https://github.com/namecoin/namecoin-core/pull/321
for the 1.2 release branch.

We need this as a first step for backporting
https://github.com/namecoin/namecoin-core/pull/322 to Xaya 1.2.
Unlike Namecoin, having this functionality in Xaya is quite important
and useful, thus we should make sure it gets included in a release as
soon as possible.
