+++ src/Makefile.test.include
@@ -95,6 +95,7 @@ BITCOIN_TESTS =\
   test/merkleblock_tests.cpp \
   test/multisig_tests.cpp \
   test/name_tests.cpp \
+  test/name_mempool_tests.cpp \
   test/net_tests.cpp \
   test/netbase_tests.cpp \
   test/pmt_tests.cpp \
