+++ src/miner.cpp
@@ -15,6 +15,7 @@
 #include <consensus/validation.h>
 #include <hash.h>
 #include <net.h>
+#include <names/main.h>
 #include <policy/feerate.h>
 #include <policy/policy.h>
 #include <pow.h>
