+++ src/validation.cpp
@@ -17,6 +17,8 @@
 #include <cuckoocache.h>
 #include <hash.h>
 #include <index/txindex.h>
+#include <names/main.h>
+#include <names/mempool.h>
 #include <policy/fees.h>
 #include <policy/policy.h>
 #include <policy/rbf.h>
